move the open deep link to manager
