+++ sdk/src/main/java/io/radar/sdk/RadarInAppMessageReceiver.kt
@@ -1,11 +1,8 @@
 package io.radar.sdk
+
 import android.app.Activity
 import android.content.Context
-import android.content.Intent
-import android.util.Log
 import android.view.View
-import androidx.core.net.toUri
-
 import io.radar.sdk.model.RadarInAppMessage
 
 /**
@@ -34,23 +31,8 @@ interface RadarInAppMessageReceiver {
      * @param inAppMessage The payload containing the message data
      */
     fun onInAppMessageButtonClicked(inAppMessage: RadarInAppMessage) {
-        val activity = activity
-        Log.d("MyInAppMessageReceiver", "called super, activity is ${activity}")
-        if (inAppMessage.button?.deepLink != null && inAppMessage.button.deepLink != "null" && inAppMessage.button.deepLink.isNotBlank() && activity != null) {
-            inAppMessage.button.deepLink.let { deepLink ->
-                try {
-                    val uri = deepLink.toUri()
-                    Radar.logger.d("Opening URL: $deepLink -> URI: $uri")
-                    val intent = Intent(Intent.ACTION_VIEW, uri)
-                    activity.startActivity(intent)
-                } catch (e: Exception) {
-                    // Handle invalid URL or no app to handle the intent
-                    Radar.logger.e("Error opening URL '$deepLink': ${e.message}")
-                }
-            }
-        } else {
-            Radar.logger.d("Button deepLink is null or 'null' string, skipping deepLink opening")
-        }
+        return 
+        
     }
 
     fun createInAppMessageView(
